feat: Integrate ConfigurationHealthReporter into ConfigurationState for improved health reporting
